+++ packages/editor/CHANGELOG.md
@@ -5,6 +5,8 @@
 - Various bug fixes.
 - Fixed an issue where the style engine generated incorrect selectors on both server and client sides.
 - Block section settings are now hidden if there are no taggable features in the section.
+- Fixed an issue where the toggle action in the block section was not working correctly.
+- Fixed an issue where the block section focus, collapse and expand actions were not being updated correctly as a unified state around the editor for all blocks.
 
 ### Development Workflow
 - GitHub workflow to enable developers to have custom PR playground configurations.
@@ -29,6 +31,7 @@
 
 ### Improvements
 - The overall design of the block settings has been improved to boost the UX (based on user feedback and user researches).
+- Enhanced Block Sections with improved state management for seamless transitions between parent and inner blocks, maintaining section states and focus mode settings during block switching.
 
 ### New Features:
 - Added Box Sizing feature to Size block section to enhance design flexibility in size-related customizations.